// src/types/partials/ValueSetCompose.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_classes so do not edit manually

import { TExtension } from '../partials/Extension';
import { TDate } from '../simpleTypes/Date';
import { TValueSetInclude } from '../partials/ValueSetInclude';

export type TValueSetCompose = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    lockedDate?: TDate;
    inactive?: Boolean;
    include: TValueSetInclude[];
    exclude?: TValueSetInclude[];
};